* feature/conversion_errors: (30 commits)
  new test set up
  shorten code branchnode
  removed cmp operators branchcand
  usage of different settings mip
  BranchNode major update
  only stab at the end
  added static GRBenv
  added new creation of child nodes
  conversion errors arctime
  use std::tie Interval
  update sttings vs code
  removed cmp templates bb
  summary python updqte
  all instances solved
  used std::tie for  cmp Instance
  used std::tie for cmp
  no heuristic iterations  when sol = 0
  cmake cleanup
  cleanup end
  conversion error branch_id
  ...

+++ include/NodeBddDumper.hpp
@@ -1,6 +1,7 @@
 #ifndef NODE_BDD_DUMPER_HPP
 #define NODE_BDD_DUMPER_HPP
 
+#include <cstddef>
 #include <unordered_set>
 #include "NodeBdd.hpp"
 #include "util/MyHashTable.hpp"
@@ -14,8 +15,8 @@
 template <typename S, typename T = NodeBdd<double>>
 class DdDumper {
     using Spec = S;
-    static int const AR = Spec::ARITY;
-    static int const headerSize = 1;
+    static size_t const AR = Spec::ARITY;
+    static int const    headerSize = 1;
 
     /* SpecNode
      * ┌────────┬────────┬────────┬─────
